move template function body into header file foregoing explicit template initialization due to compile errors on clang
